feat:注意private Integer pid;，只能是这个类型，若改为Long，则首页白屏，无法展示

+++ bubble-api/bubble-api-oa/src/main/java/com/bubblecloud/oa/api/vo/MenuTreeNodeVO.java
@@ -26,7 +26,7 @@ public class MenuTreeNodeVO {
 	private Long id;
 
 	@Schema(description = "父菜单ID")
-	private Long pid;
+	private Integer pid;
 
 	@Schema(description = "菜单名称")
 	private String menuName;
